requirements.txt
streamlit
qrcode
Pillow 